wip(removal): decouple slices in flight (checkpoint)
