+++ crates/stella-plugin/src/lib.rs
@@ -91,6 +91,7 @@
 //! capability is never performed — however eagerly the plugin's own process
 //! would answer or ask for one.
 
+mod candidate_grant;
 mod consent;
 mod driver;
 mod error;
@@ -113,6 +114,16 @@ mod wire;
 pub mod wire_corpus;
 mod wrapper;
 
+// The host-side data and logic behind a candidate grant (removal census for
+// `stella-pipeline`, `docs/spec/pipeline-as-plugins.md` §7 slice 1, §1.5.6) —
+// see the module doc for why the process-outcome types, the test-command
+// parser, the tamper comparator, and the path fence all live beside the wire
+// types they build against.
+pub use candidate_grant::{
+    ArtifactIdentity, ArtifactKind, CmdKind, CmdOutcome, HOST_TREE_HANDLE, TestInvocation,
+    TestInvocationError, canonical_root, fence, host_tree_grant, parse_test_invocation,
+    resolve_in_root, test_plan, witness_identity_matches,
+};
 pub use consent::{Capability, RiskLevel, consent_text, highest_risk};
 pub use driver::{
     DriveNext, DrivePoint, DriveRequest, DriveResponse, DriveSession, DriverCall,
